UtopiaBot-Jr/requirements.txt
# TikTok TTS Bot - Python Requirements
# Version 1.0 - July 27, 2025

# Core TikTok Live Integration
TikTokLive>=6.6.0

# Google Cloud Text-to-Speech
google-cloud-texttospeech>=2.16.0

# Audio Playback
playsound>=1.3.0

# Text Processing
emoji>=2.8.0

# GUI Framework (Built-in with Python)
# tkinter - No installation needed (part of Python standard library)

# HTTP Requests for Connection Testing
requests>=2.31.0

# Threading and Async Support (Built-in with Python)
# asyncio - No installation needed (part of Python standard library)
# threading - No installation needed (part of Python standard library)

# File and System Operations (Built-in with Python)
# os - No installation needed (part of Python standard library)
# time - No installation needed (part of Python standard library)
# datetime - No installation needed (part of Python standard library)
# queue - No installation needed (part of Python standard library)
# random - No installation needed (part of Python standard library)
# logging - No installation needed (part of Python standard library)

# Optional: For development and testing
# pytest>=7.4.0
# pylint>=2.17.0
# black>=23.7.0

# Installation Instructions:
# 1. Install all requirements: pip install -r requirements.txt
# 2. Or install individually:
#    pip install TikTokLive google-cloud-texttospeech playsound emoji requests

# Python Version Requirement:
# Python 3.8 or higher required
